Use a pool for hub requests
Testing suggests that sending a second PUT while the first is still
pending can confuse the hub, and result in just one of the commands
actually happening, while updating the hub status of both.

Work around this by using a pool for our requests with a maxSockets
of 1, so that we only ever send one at a time.

For safety use this for GETs too.