Separate first and subsequent setPosition delays
The first delay is there for scene coalescing, so we don't send two
different PUTs for the same shadeId when setting scenes involving
multiple positions (e.g. duette).

The second delay is there in case there's an issue with two PUTs
within a certain time period. Right now we just avoid overlapping
them, so this is set to the same delay as above, but if I find I'm
still getting dropped shade movements from the PowerView hub, I can
increase this speculatively.