Separate out hub methods into their own file
This keeps the whole request API pretty clean, since we can hide the
err/response/statusCode stuff and just deal in the blocks of data we
understand.

Make one major change in the process, it turns out that if you want to
just set position2, you do NOT send:

  { posKind2: 2, position2: xxx }

You actually send:

  { posKind1: 2, position1: xxx }

So introduce a putShadePosition API that handles the merge smartly that
way, adding a new position value or overwriting an existing one.