Add the ability to refresh the shade positions
When we send ?refresh=true, the hub actually goes and talks to the
shade to figure out where it is. We have to queue these up though
because otherwise the hub likes to crash.

Right now I'm about 50/50 about whether this is the right thing to do
for CurrentPosition.get, but all I need to do to swap it is flip a
boolean, so I'll keep it for now.