Add support for Silhouette shades
These have just the single accessory, but that accessory gains two
extra characteristics for Horizontal Tilt.

Setting this is a bit funky since it's done with posKindN: 3, which
replaces posKindN: 1 as long as the shades are closed. Setting it
also happens to close the shades, and likewise setting the bottom will
flatten the vanes.