Refresh shades when position is missing
This may end up being the best compromise between refreshing all the
time, and none of the time.

When we adjust just one of the two Duette accessories, the hub stops
returning the position of the other, so we get null/undefined; use that
as a signal to try again with ?refresh=true so we restore the full data.