Using shade id, instead of shade name to generate the accessory UUID
It's possible and completely reasonable for shades to use the same name in different rooms. The UUIDGen seems to generate the same UUID given the same input string. This means when multiple shades have the same name, we end up with UUID collisions and homebridge will generate an error. Instead, if we use the shade id (which is unique), we can generate a unique accessory UUID and prevent the error.